remove MapperError in favor of TranspilerError
