+++ qiskit/mapper/__init__.py
@@ -10,4 +10,4 @@
 from .compiling import two_qubit_kak, euler_angles_1q
 from .coupling import CouplingMap
 from .layout import Layout
-from .exceptions import CouplingError, MapperError
+from .exceptions import CouplingError, LayoutError
